cmp_http.c: fix memory leak in set_http_path()
git-svn-id: svn://svn.code.sf.net/p/cmpforopenssl/code@280 8c94e50a-443e-0410-be32-d2e341db1201

+++ src/openssl-client/cmpclient.c
@@ -1030,7 +1030,7 @@ int getHttpProxy( char **name, int *port) {
   maxlen = strlen(proxy) + 1;
   *name = calloc(1, maxlen);
 
-  snprintf(format, 32, "http //%%%us %%d", maxlen);
+  snprintf(format, 32, "http //%%%us %%d", (unsigned int) maxlen);
 
   if( (sscanf( proxy, format, *name, port) < 1)) {
     /* maybe it is set without leading http:// */
